Update font, add thumbnails & reorder landing page
Tried two more fonts to increase readability and style. The second
option (Love Me) proved to be the best option for now. Added the
thumbnail images to each video. Reordered the landing page, so that the
banner text is now the hero section, and the
about description is below the how it works section. This prevents
uneccessary repetition and improves the style.